change: directory structure and deleting some existing files
