--- cmd/main/test_one.go
@@ -1,7 +1,0 @@
-package main
-
-import "fmt"
-
-func testing_one() {
-	fmt.Println("Hello testing")
-}
